+++ app/Http/Controllers/itemController.php
@@ -11,7 +11,7 @@ class itemController extends Controller
 {
     public function __construct(Item $item){
       $this->item = $item;
-      //$this->middleware('auth:api', ['except' => ['index']]);
+      $this->middleware('auth:api', ['except' => ['index']]);
     }
 
     /**
@@ -42,6 +42,10 @@ class itemController extends Controller
      */
     public function store(Request $request)
     {
+      if ($request->header('admin') != "true") {
+        return response()->json(["Error" => "not worthy"], 401);
+      }
+
       $newStuff = [
         "category_id" => $request->category_id,
         "name" => $request->name,
@@ -57,7 +61,7 @@ class itemController extends Controller
         return response()->json($array);
       }
       catch(QueryException $a){
-        return response()->json(["Error" => "it screwed up"], 404);
+        return response()->json(["Error" => $a], 404);
       }
     }
 
@@ -119,6 +123,9 @@ class itemController extends Controller
       if(isset($request->stock)) $newStuff["stock"] = $request->stock;
 
       try{
+        if ($request->header('admin') != "true") {
+          return response()->json(["Error" => "not worthy"], 401);
+        }
         $data = $this->item->where("id",$id)->update($newStuff);
         //return response()->json($data);
       }
@@ -133,9 +140,12 @@ class itemController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
       try{
+        if ($request->header('admin') != "true") {
+          return response()->json(["Error" => "not worthy"], 401);
+        }
         $data = $this->item->where("id",$id)->delete();
       }
       catch(QueryException $a){
